+++ gaios/Sendflow/Flow/Failure/SendFailureViewModel.swift
@@ -3,13 +3,17 @@ import UIKit
 
 @MainActor
 final class SendFailureViewModel: Sendable {
-    
+
     let delegate: SendFailureViewModelDelegate?
     let error: Error
+    let hideErrors: Bool
 
-    internal init(delegate: (any SendFailureViewModelDelegate)? = nil, error: any Error) {
+    internal init(delegate: (any SendFailureViewModelDelegate)? = nil,
+                  error: any Error,
+                  hideErrors: Bool) {
         self.delegate = delegate
         self.error = error
+        self.hideErrors = hideErrors
     }
 
     func onDismiss() {
@@ -22,5 +26,8 @@ final class SendFailureViewModel: Sendable {
 
     func onSupport() {
         delegate?.sendFailureViewModelDidSupport(self, error: error)
+    }
+    func onOk() {
+        delegate?.sendFailureViewModelOk(self)
     }
 }